+++ skeleton lab2/fbputchar_.h
@@ -6,9 +6,9 @@
 #define FBOPEN_VSCREENINFO -3  /* Couldn't read the variable info */
 #define FBOPEN_MMAP -4         /* Couldn't mmap the framebuffer memory */
 #define FBOPEN_BPP -5          /* Unexpected bits-per-pixel */
-
+#include <stdbool.h>
 extern int fbopen(void);
-extern void fbputchar(char, int, int);
+extern void fbputchar(char, int, int, bool);
 extern void fbputs(const char *, int, int);
 
 #endif
